Рефакторит код, заполняет даты нового события по умолчанию
