+++ src/presenter/waypoint-new-presenter.js
@@ -1,7 +1,7 @@
 import AddFormView from '../view/add-form-view.js';
 import { isEscapeKey } from '../utils.js';
 import { render, remove, RenderPosition } from '../framework/render.js';
-import { UserAction, UpdateType, DEFAULT_WAY_POINT } from '../const.js';
+import { UserAction, UpdateType, BLANK_WAYPOINT } from '../const.js';
 
 export default class WayPointNewPresenter {
 
@@ -24,7 +24,7 @@ export default class WayPointNewPresenter {
       return;
     }
 
-    this.#wayPointAddComponent = new AddFormView(DEFAULT_WAY_POINT,
+    this.#wayPointAddComponent = new AddFormView(BLANK_WAYPOINT,
       this.#wayPointsModel.destinations,
       this.#wayPointsModel.offers);
     this.#wayPointAddComponent.setFormSubmitHandler(this.#handleFormSubmit);
