+++ src/waypoints-api-service.js
@@ -30,9 +30,7 @@ export default class WayPointsApiService extends ApiService {
       body: JSON.stringify(this.#adaptToServer(wayPoint)),
       headers: new Headers({ 'Content-Type': 'application/json' }),
     });
-
     const parsedResponse = await ApiService.parseResponse(response);
-
     return parsedResponse;
   };
